Standarized Weapon Logic into Weapon
