+++ business/weapons/weapon_factory.py
@@ -7,21 +7,21 @@ class WeaponFactory:
     def create_weapon(weapon_type: str):
         if weapon_type == "Auto_Joker":
             bullet_name = "Joker_Bullet"
-            bullet_cooldown = 5000
+            bullet_cooldown = 10000
             bullet_speed = 20
             return AutoAimGun(weapon_type, bullet_name, bullet_cooldown, bullet_speed)
         if weapon_type == "Manual_Gun":
             bullet_name = "Bullet"
-            bullet_cooldown = 500
+            bullet_cooldown = 200
             bullet_speed = 20
             return ManualGun(weapon_type, bullet_name, bullet_cooldown, bullet_speed)
         if weapon_type == "Manual_Joker":
             bullet_name = "Joker_Bullet"
-            bullet_cooldown = 2000
+            bullet_cooldown = 1300
             bullet_speed = 20
             return ManualGun(weapon_type, bullet_name, bullet_cooldown, bullet_speed)
         if weapon_type == "The_Mega_Ice":
             bullet_name = "Ice_Bullet"
-            bullet_cooldown = 2000
+            bullet_cooldown = 2705
             bullet_speed = 20
             return ManualGun(weapon_type, bullet_name, bullet_cooldown, bullet_speed)
